requirements.txt
# Core dependencies
langchain>=0.3.16,<0.4.0
langchain_openai
langchain_chroma
langchain-community>=0.3.16,<0.4.0
chromadb>=0.4.14
flask==3.0.1
flask_cors==3.0.10
google.generativeai
pydantic>=2.7.4,<3.0.0
streamlit
# PDF Processing
pypdf==4.2.0

# ML/AI Dependencies (with CPU-only versions)
sentence_transformers==2.3.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu